% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age.old.nsr.maltman}
\alias{age.old.nsr.maltman}
\title{ages-by-nsr}
\format{
An object of class \code{list} of length 2.
}
\usage{
age.old.nsr.maltman
}
\description{
ages-by-nsr
}
\keyword{datasets}
