% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_landcover.R
\name{clean_landcover}
\alias{clean_landcover}
\title{Clean landcover data to match ABMI species model coefficients}
\usage{
clean_landcover(
  data.in,
  landscape.lookup = landcover.coef.lookup,
  type = "Vegetation",
  burn.cc = "Harvest",
  class.in = "ID",
  class.out = "COEF"
)
}
\arguments{
\item{data.in}{Output from \code{make_landcover_wide()}}

\item{landscape.lookup}{Defines the lookup table used for aggregating the landcover data. Default is provided lookup}

\item{type}{Character; Defines if the landcover data is "Soil" or "Vegetation"}

\item{burn.cc}{character; If "Harvest", harvests that have burned are treated as if fires reset orign year. If "Vegetation", it assigns it to the closest native vegetation.}

\item{class.in}{Input column that matches classes from make_landcover_wide}

\item{class.out}{Output classes that match the coefficient}
}
\description{
Create veg, hf, and soil summaries from GIS back-filled data in wide format that align with the ABMI species models
}
\examples{
\dontrun{
landcover.out <- clean_landcover(data.in = d.wide,
                                 landscape.lookup = landcover.coef.lookup,
                                 type = "Vegetation",
                                 burn.cc = "Harvest",
                                 class.in = "ID",
                                 class.out = "COEF")
}

}
