% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.fill_in_0ages}
\alias{.fill_in_0ages}
\title{Fill in 0 ages}
\usage{
.fill_in_0ages(x, NSR, ages.list, age.data = "Maltman.Old")
}
\arguments{
\item{x}{Input}

\item{NSR}{NSR column}

\item{ages.list}{Ages list object}

\item{age.data}{Default; Maltman.old. Defines which age information will be used for unknown stands (ABMI.Young; Maltman.Young; ABMI.Old; Maltman.Old)}
}
\description{
This function redistributes unknown ages in wide form summaries
}
