% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.make_char2fact}
\alias{.make_char2fact}
\title{Internal functions}
\usage{
.make_char2fact(x)
}
\arguments{
\item{x}{Input}
}
\description{
Make factors from characters in SQL tables
}
\details{
Character to Factor
}
