% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{footprint.lookup}
\alias{footprint.lookup}
\title{hfi-lookup}
\format{
An object of class \code{data.frame} with 229 rows and 11 columns.
}
\usage{
footprint.lookup
}
\description{
hfi-lookup
}
\keyword{datasets}
