% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{harvest.lookup}
\alias{harvest.lookup}
\title{harvest-areas-lookup}
\format{
An object of class \code{data.frame} with 20 rows and 3 columns.
}
\usage{
harvest.lookup
}
\description{
harvest-areas-lookup
}
\keyword{datasets}
