% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_unknown_ages.R
\name{impute_unknown_ages}
\alias{impute_unknown_ages}
\title{impute_unknown_ages}
\usage{
impute_unknown_ages(
  landcover,
  ages.list,
  col.baseyear,
  decadal.fix = TRUE,
  origin.fix = TRUE,
  unknown.fix = FALSE,
  age.data = "Maltman.Old",
  ver.id = "V7.0"
)
}
\arguments{
\item{landcover}{Input}

\item{ages.list}{Ages list object}

\item{col.baseyear}{Base year for surveys or HF inventory. Used to patch disturbances}

\item{decadal.fix}{Logical; Determines if decades prior to 1960 should be smoothed}

\item{origin.fix}{Logical; Determines if origin years should be assigned based off Maltman et al when possible}

\item{unknown.fix}{Logical; Determines if we assign origin years to remaining unknown stands. Use TRUE for reporting, FALSE for site summaries}

\item{age.data}{Default; Maltman.old. Defines which age information will be used for unknown stands (ABMI.Old; Maltman.Old)}

\item{ver.id}{Character; defines which version of the backfill is being processed. Supports V6.1 and V7.0}
}
\description{
This function imputes unknown ages into the long form. This is used for tracking changes over time
}
