% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kgrid}
\alias{kgrid}
\title{unknown-ages-lookup}
\format{
An object of class \code{data.frame} with 664762 rows and 4 columns.
}
\usage{
kgrid
}
\description{
unknown-ages-lookup
}
\keyword{datasets}
