% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landcover.coef.lookup}
\alias{landcover.coef.lookup}
\title{landcover-coefficient-lookup}
\format{
An object of class \code{list} of length 2.
}
\usage{
landcover.coef.lookup
}
\description{
landcover-coefficient-lookup
}
\keyword{datasets}
