% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_veg_long.R
\name{make_landcover_long}
\alias{make_landcover_long}
\title{Process input data into long summary}
\usage{
make_landcover_long(
  landcover,
  col.label,
  col.baseyear = NULL,
  col.hfyear = "YEAR",
  col.veg = "Combined_ChgByCWCS",
  col.soil = "Soil_Type_1",
  hf.fine = TRUE,
  burn.cc = TRUE,
  unround = FALSE,
  age.correction = FALSE,
  ver.id = "V7.0"
)
}
\arguments{
\item{landcover}{Data frame of summary read in using \code{read_summary()}}

\item{col.label}{Unique ID (can all be same value); e.g., "Site_ID"}

\item{col.baseyear}{Base year for surveys or HF inventory. Used to calculate years since last disturbance. Use numeric when it is the same for each record, character to indicate field when it varies by record.}

\item{col.hfyear}{HF year column; default is "YEAR"}

\item{col.veg}{Vegetation column; default is "Combined_ChgByCWCS"}

\item{col.soil}{Soil column; default is "Soil_Type_1"}

\item{hf.fine}{Logical; defaults to TRUE. Decides if coarse of fine footprint levels should be used.}

\item{burn.cc}{Logical; defaults to TRUE. Tracks harvest areas that have been new and their new disturbance date (TRUE) or maintains cutblock status with harvest origin year age (FALSE).}

\item{unround}{Logical; whether to unround rounded origin year values pre-2000}

\item{age.correction}{Age correction column; applies built in age correction old forests. Default of FALSE means no age correction occurs.}

\item{ver.id}{Character; defines which version of the backfill is being processed. Supports V6.1 and V7.0}
}
\description{
Create veg, hf, and soil summaries from GIS back-filled data in long format
}
\examples{
\dontrun{
d.long <- make_veghf_long(landcover = df,
                          col.label = "Region",
                          col.baseyear = 2017,
                          col.hfyear = "YEAR",
                          col.veg = "Combined_ChgByCWCS",
                          col.soil = "Soil_Type_1",
                          hf_fine = TRUE,
                          burn.cc = TRUE,
                          unround = FALSE,
                          age.correction = FALSE,
                          ver.id = "V7.0")
}

}
