% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_veg_wide.R
\name{make_landcover_wide}
\alias{make_landcover_wide}
\title{Process long summary into wide summary}
\usage{
make_landcover_wide(
  long.output,
  col.label,
  col.area = "Shape_Area",
  hf.fine = TRUE,
  tol = 0,
  sparse = TRUE,
  assign.unknown.ages = TRUE,
  age.data = "Maltman.Old",
  ver.id = "V7.0",
  rm0 = TRUE
)
}
\arguments{
\item{long.output}{Output from \code{make_veghf_long()}}

\item{col.label}{Unique ID (can all be same value); e.g., "Site_ID"}

\item{col.area}{Column of polygon area; defaults to "Shape_Area"}

\item{hf.fine}{Logical; defaults to TRUE}

\item{tol}{Tolerance level for excluding unknown aged harvest areas}

\item{sparse}{Logical; defaults to TRUE}

\item{assign.unknown.ages}{Logical; defaults to TRUE, assigns unknown ages based on the proportions in each natural subregion}

\item{age.data}{Default; Maltman.old. Defines which age information will be used for unknown stands (ABMI.Young; Maltman.Young; ABMI.Old; Maltman.Old)}

\item{ver.id}{Character; defines which version of the backfill is being processed. Supports V6.1 and V7.0}

\item{rm0}{Logical; defines if the original column for unknown ages is removed after age assignment}
}
\description{
Create veg, hf, and soil summaries from GIS back-filled data in wide format
}
\examples{
\dontrun{
d.wide <- make_veghf_wide(long.output = d.long,
                          col.label = "Site_ID",
                          col.area = "Shape_Area",
                          hf.fine = TRUE,
                          tol = 0,
                          sparse = TRUE,
                          assign.unknown.ages = TRUE,
                          age.data = "Maltman.Old",
                          ver.id = "V7.0",
                          rm0 = TRUE)
}

}
