% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_veg_long.R
\name{make_veghf_long}
\alias{make_veghf_long}
\title{Process input data into long summary}
\usage{
make_veghf_long(
  d,
  col.label,
  col.baseyear = NULL,
  col.hfyear = "YEAR",
  col.veg = "Combined_ChgByCWCS",
  col.soil = "Soil_Type_1",
  hf_fine = TRUE,
  unround = FALSE
)
}
\arguments{
\item{d}{Data frame of summary read in using \code{read_summary()}}

\item{col.label}{Unique ID (can all be same value); e.g., "Site_ID"}

\item{col.baseyear}{Base year for surveys or HF inventory. Used to calculate years since last disturbance. Use numeric when it is the same for each record, character to indicate field when it varies by record.}

\item{col.hfyear}{HF year column; default is "YEAR"}

\item{col.veg}{Vegetation column; default is "Combined_ChgByCWCS"}

\item{col.soil}{Soil column; default is "Soil_Type_1"}

\item{hf_fine}{Logical; defaults to TRUE}

\item{unround}{Logical; whether to unround rounded origin year values pre-2000}
}
\description{
Create veg, hf, and soil summaries from GIS back-filled data in long format
}
\examples{
\dontrun{
d_long <- make_veghf_long(d = df,
                          col.label = "Region",
                          col.baseyear = 2017,
                          col.hfyear = "YEAR",
                          col.veg = "Combined_ChgByCWCS",
                          col.soil = "Soil_Type_1",
                          hf_fine = TRUE,
                          unround = FALSE)
}

}
