% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_veg_wide.R
\name{make_veghf_wide}
\alias{make_veghf_wide}
\title{Process long summary into wide summary}
\usage{
make_veghf_wide(
  d,
  long_output,
  col.label,
  col.area = "Shape_Area",
  hf_fine = TRUE,
  tol = 0,
  sparse = TRUE
)
}
\arguments{
\item{d}{Data frame of summary read in using \code{read_summary()}}

\item{long_output}{Output from \code{make_veghf_long()}}

\item{col.label}{Unique ID (can all be same value); e.g., "Site_ID"}

\item{col.area}{Column of polygon area; defaults to "Shape_Area"}

\item{hf_fine}{Logical; defaults to TRUE}

\item{tol}{Tolerance level for excluding unknown aged harvest areas}

\item{sparse}{Logical; defaults to TRUE}
}
\description{
Create veg, hf, and soil summaries from GIS back-filled data in wide format
}
\examples{
\dontrun{
d_wide <- make_veghf_wide(d = df, long_output = d_long, col.label = "Site_ID")
}

}
