% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nsr.age.proportions}
\alias{nsr.age.proportions}
\title{ages-by-nsr}
\format{
An object of class \code{list} of length 2.
}
\usage{
nsr.age.proportions
}
\description{
ages-by-nsr
}
\keyword{datasets}
