% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_summary.R
\name{read_summary}
\alias{read_summary}
\title{Read in summary data}
\usage{
read_summary(summary.path, table = NULL, col.subset = NULL, val.subset = NULL)
}
\arguments{
\item{summary.path}{Path to summary file (can be .csv or .sqlite)}

\item{table}{Name of table of interest for SQLite database (ignored for csv files)}

\item{col.subset}{Optional; column name to be used for subsetting}

\item{val.subset}{Values in \code{col.subset} to keep. Can be single value or character vector (e.g., c("NE", "NW", "SE", "SW"))}
}
\description{
Read in summary data provided by the Geospatial Centre
}
