% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soil.lookup}
\alias{soil.lookup}
\title{soil-lookup}
\format{
An object of class \code{data.frame} with 28 rows and 2 columns.
}
\usage{
soil.lookup
}
\description{
soil-lookup
}
\keyword{datasets}
