% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vegetation.lookup}
\alias{vegetation.lookup}
\title{vegetation-lookup}
\format{
An object of class \code{data.frame} with 59 rows and 6 columns.
}
\usage{
vegetation.lookup
}
\description{
vegetation-lookup
}
\keyword{datasets}
