% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.fill_in_0ages}
\alias{.fill_in_0ages}
\title{Fill in 0 ages}
\usage{
.fill_in_0ages(x, NSR, ages_list, rm0 = TRUE)
}
\arguments{
\item{x}{Input}

\item{NSR}{NSR column}

\item{ages_list}{Ages list object}

\item{rm0}{Logical}
}
\description{
This function redistributes unknown ages in wide form summaries
}
