% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veghfsoil.R
\docType{package}
\name{veghfsoil}
\alias{veghfsoil}
\alias{veghfsoil-package}
\title{veghfsoil}
\description{
Take GIS site information and produce Veg+HF+Soil summaries for species modeling.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ABbiodiversity/veghfsoil}
  \item Report bugs at \url{https://github.com/ABbiodiversity/veghfsoil/issues}
}

}
\author{
Marcus Becker \email{mabecker89@ualberta.ca}
}
